Resolve quesito 9

+++ services/salesServices.js
@@ -1,15 +1,16 @@
 const salesModel = require('../models/salesModel');
 const salesSchema = require ('../Schema/salesSchema');
+const { atualizarEstoque, atualizarEstoqueDelete } = require('../utils/atualizarEstoque');
 
 const ZERO_MODIFIED = 0;
 
 const insert = async (itens) => {
   const validations = await salesSchema.validate(itens);
-
   if (validations) return validations;
 
-  const data = await salesModel.insert(itens);
+  atualizarEstoque(itens);
 
+  const data = await salesModel.insert(itens);
   if(!data) return null;
 
   return { status: 200, data };
@@ -32,20 +33,21 @@ const getAll = async () => {
 
 const updateByID = async (id, itensSold) => {
   const validations = await salesSchema.validate(itensSold);
-
   if (validations) return validations;
 
   const data = await salesModel.updateByID(id, itensSold);
-
   if(data.modifiedCount === ZERO_MODIFIED) return null;
 
   return { status: 200, _id: id, itensSold };
 };
 
 const deleteByID = async (id) => {
-  const data = await salesModel.deleteByID(id);
+  const sale = await findById(id);
+  if(!sale) return null;
+  await atualizarEstoqueDelete(sale.data.itensSold);
 
-  if(data && data.deletedCount && data.deletedCount === 1) return { status: 200, data };
+  const resp = await salesModel.deleteByID(id);
+  if(resp && resp.deletedCount && resp.deletedCount === 1) return { status: 200, resp };
 
   return null;
 };
